pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cropgen"
version = "0.1.0"
description = "First step in the generation of the dataset for fine-tuning an LLM in vision tasks."
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.2.0",
    "label-studio-sdk>=2.0.19",
    "openpyxl>=3.1.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "shapely>=2.1.2",
    "tqdm>=4.67.3",
]

[tool.setuptools.packages.find]
where = ["src"]